refactor: rename fast => fastest
